+++ module/actor/actor-beast-sheet.js
@@ -97,7 +97,6 @@ export class ArM5eBeastActorSheet extends ArM5eActorSheet {
       case "ability":
       case "diaryEntry":
       case "powerFamiliar":
-      case "magicItem":
       case "personalityTrait":
       case "reputation":
       case "magicalEffect":