+++ module/config.js
@@ -2679,6 +2679,8 @@ ARM5E.lab.enchantment.effectUses = {
   10: "Unlimited"
 };
 
+ARM5E.lab.enchantment.enchantableTypes = ["item", "weapon", "armor", "book"];
+
 ARM5E.generic = {};
 ARM5E.generic.sourcesTypes = {
   custom: {